+++ NEWS.md
@@ -1,7 +1,7 @@
 # ICAMS 2.2.5 
 ## Changed
 * Updated function `CreateOneColSBSMatrix` to throw a message instead of an error when there are SBS variant whose reference base in ref.genome does not match the reference base in the VCF file.
-* Updated function `MakeVCFDBSdf` to use column `read.depth` instead of `TUMOR` from original VCF to calculate the read depth for DBS(From merged SBSs).
+* Updated function `MakeVCFDBSdf` to inherit column information from original SBS VCF.
 * Changed the words in legend for DBS144 plot from "Transcribed", "Untranscribed" to "Transcribed strand" and "Untranscribed strand".
 
 <br/>